requirements.txt
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-cors-headers==3.7.0
django-rest-swagger==2.1.2
djangorestframework==3.9.1
idna==2.8
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==1.4.6
openapi-codec==1.3.2
psycopg2==2.9.1
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.22.0
simplejson==3.17.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
validate-email==1.3
